requirements.txt
spotdl==4.3.0
Flask==3.1.1
waitress==3.0.2
python-dotenv==1.0.1
gunicorn>=21.2
yt-dlp>=2024.03.10
mutagen>=1.47
